requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
colorama==0.4.6
cryptography==41.0.4
et-xmlfile==1.1.0
h11==0.14.0
idna==3.3
numpy==1.24.3
outcome==1.2.0
pandas==2.0.1
psutil==5.9.5
pycparser==2.21
pycryptodome==3.19.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
selenium==4.13.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.1
tqdm==4.65.0
trio==0.22.2
trio-websocket==0.11.1
tzdata==2023.3
urllib3==1.26.7
wsproto==1.2.0
